+++ Dockerfile
@@ -1,7 +1,7 @@
 # vim:set ft=dockerfile:
 FROM continuumio/miniconda3
-MAINTAINER https://github.com/roocs/roocs-wps-demo
-LABEL Description="roocs wps demo WPS" Vendor="Birdhouse" Version="0.1.0"
+MAINTAINER https://github.com/roocs/rook
+LABEL Description="rook WPS" Vendor="roocs" Version="0.1.0"
 
 # Update Debian system
 RUN apt-get update && apt-get install -y \
@@ -25,9 +25,9 @@ RUN ["/bin/bash", "-c", "source activate wps && python setup.py develop"]
 # Start WPS service on port 5000 on 0.0.0.0
 EXPOSE 5000
 ENTRYPOINT ["/bin/bash", "-c"]
-CMD ["source activate wps && exec roocswps start -b 0.0.0.0 -c /opt/wps/etc/demo.cfg"]
+CMD ["source activate wps && exec rook start -b 0.0.0.0 -c /opt/wps/etc/demo.cfg"]
 
-# docker build -t roocs/roocs-wps-demo .
-# docker run -p 5000:5000 roocs/roocs-wps-demo
+# docker build -t roocs/rook .
+# docker run -p 5000:5000 roocs/rook
 # http://localhost:5000/wps?request=GetCapabilities&service=WPS
 # http://localhost:5000/wps?request=DescribeProcess&service=WPS&identifier=all&version=1.0.0
