+++ etc/demo.cfg
@@ -2,5 +2,5 @@
 level = DEBUG
 
 [data]
-cmip5_archive_root = /data/roocs-wps-demo/mini-esgf-data/test_data/badc/cmip5/data
-cordex_archive_root = /data/roocs-wps-demo/mini-esgf-data/test_data/group_workspaces/jasmin2/cp4cds1/data/c3s-cordex
+cmip5_archive_root = /data/rook/mini-esgf-data/test_data/badc/cmip5/data
+cordex_archive_root = /data/rook/mini-esgf-data/test_data/group_workspaces/jasmin2/cp4cds1/data/c3s-cordex
