+++ requirements.txt
@@ -1,5 +1,8 @@
-pywps>=4.2.0<4.3
+pywps>=4.2.0
 jinja2
 click
 psutil
 networkx
+xarray>=0.15
+dask[complete]
+netcdf4
