+++ rook/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Top-level package for roocs wps demo."""
+"""Top-level package for rook."""
 
 from .__version__ import __author__, __email__, __version__  # noqa: F401
