+++ rook/default.cfg
@@ -1,10 +1,10 @@
 [metadata:main]
-identification_title = roocs wps demo
-identification_abstract = A demostrator for a WPS service for roocs.
+identification_title = rook
+identification_abstract = A WPS service for roocs.
 identification_keywords = PyWPS, WPS, OGC, processing, birdhouse, roocs, demo, cp4cds, copernicus, ecmwf
 identification_keywords_type = theme
-provider_name = roocs wps demo
-provider_url=http://roocs-wps-demo.readthedocs.org/en/latest/
+provider_name = rook
+provider_url=http://rook.readthedocs.org/en/latest/
 
 [server]
 url = http://localhost:5000/wps
@@ -16,7 +16,7 @@ parallelprocesses = 2
 
 [logging]
 level = INFO
-file = roocswps.log
+file = rook.log
 format = %(asctime)s] [%(levelname)s] line=%(lineno)s module=%(module)s %(message)s
 
 [data]
