+++ rook/exceptions.py
