+++ rook/operator.py
