+++ setup.cfg
@@ -6,7 +6,7 @@ tag = True
 [metadata]
 description-file = README.rst
 
-[bumpversion:file:roocswps/__version__.py]
+[bumpversion:file:rook/__version__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
